move Coords to uwot namespace
